LayoutTests/http/tests/security/drag-drop-local-file-expected.txt

CONSOLE MESSAGE: line 26: Not allowed to drag local resource: foobar
ALERT: PASS
Dragme 
